+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ownCloud Music Change Log
   * Switched to Chromium engine.
   * Track progress bar is integrated.
   * Volume slider is integrated.
+  * Album art is downloaded correctly.
 
 1.2 - February 12th, 2017
 -------------------------